Replace $path_info var
`$_SERVER['PATH_INFO']` isn't set by default on many servers. For instance, the existing code would not work on NGINX by default.

There is no 100% perfect way to do it, really, but this way is easy and clean.

Replaced the direct call to the `$_SERVER` var with code that rebuilds the request URL then parses it for the path.